api/requirements.txt
click==7.0
google==2.0.3
googleapis-common-protos==1.51.0
grpcio==1.26.0
grpcio-tools==1.26.0
grpcio-reflection==1.26.0
protobuf==3.11.2
protobuf-compiler==1.0.20